erased contacting method (read descritption for more detail)
The way I first wanted it to work is that you can fill out a form and then send it to my email. But due to security risks mainly because the whole website is open source online on github this doesn't really work. So I will just put my email there in the future :)